requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
akshare==1.14.1
exchange_calendars==4.5.4
pandas_ta==0.3.14b0
pymongo